refactor: emoji 관련 로직을 createPost의 엔티티로 이동하기
