+++ src/entities/createPost/libs/utils.ts
